% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{addInactiveConcepts}
\alias{addInactiveConcepts}
\title{Add inactive concepts to a SNOMEDcodelist or SNOMEDconcept vector}
\usage{
addInactiveConcepts(x, provenance = 0:3, SNOMED = getSNOMED())
}
\arguments{
\item{x}{SNOMEDcodelist or SNOMEDconcept object}

\item{provenance}{vector of provenance values to use}

\item{SNOMED}{SNOMED environment containing HISTORY and QUERY tables}
}
\value{
SNOMEDcodelist or SNOMEDconcept with linked inactive concepts included
}
\description{
Adds inactive conceptsDisplays a SNOMEDcodelist on screen, including metadata.
Truncates term descriptions in order to fit within the line width.
}
\seealso{
Other SNOMEDcodelist functions: 
\code{\link{SNOMEDcodelist}()},
\code{\link{expandSNOMED}()},
\code{\link{export}()},
\code{\link{is.SNOMEDcodelist}()},
\code{\link{print.SNOMEDcodelist}()}
}
\concept{SNOMEDcodelist functions}
