% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{c.SNOMEDconcept}
\alias{c.SNOMEDconcept}
\title{Concatenate vectors of SNOMED CT concepts}
\usage{
\method{c}{SNOMEDconcept}(...)
}
\arguments{
\item{...}{SNOMEDconcept vectors}
}
\value{
concatenation of vectors
}
\description{
SNOMEDconcept is an S3 class for vectors of SNOMED concept IDs
as 64-bit integers. This function concatenates two or more
SNOMEDconcept vectors.
}
\examples{
hf <- SNOMEDconcept('Heart failure', SNOMED = sampleSNOMED())
hf2 <- c(hf, hf)
}
\seealso{
Other SNOMEDconcept functions: 
\code{\link{as.data.frame.SNOMEDconcept}()},
\code{\link{is.SNOMEDconcept}()},
\code{\link{print.SNOMEDconcept}()},
\code{\link{union.SNOMEDconcept}()},
\code{\link{unique.SNOMEDconcept}()}
}
\concept{SNOMEDconcept functions}
