% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{checkConcepts}
\alias{checkConcepts}
\title{Check that concept IDs are in the correct format}
\usage{
checkConcepts(conceptIds)
}
\arguments{
\item{conceptIds}{character or integer64 vector}
}
\value{
conceptIds in integer64 format
}
\description{
Checks if a vector of conceptIds of type integer64, and 
converts them to integer64 if necessary.
}
\examples{
checkConcepts('900000000000003001')
}
