% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{conceptId}
\alias{conceptId}
\title{Returns the SNOMED CT concept IDs for a set of terms}
\usage{
conceptId(
  terms,
  active_only = TRUE,
  exact_match = TRUE,
  unique = TRUE,
  SNOMED = get("SNOMED", envir = globalenv())
)
}
\arguments{
\item{terms}{character vector of terms to match}

\item{active_only}{whether or not to include inactive concepts}

\item{exact_match}{if TRUE, only an exact (case sensitive)
match is performed. If FALSE, a regular expression match
is performed.}

\item{unique}{whether to include no more than one instance of each
SNOMED CT concept}

\item{SNOMED}{environment containing SNOMED dictionary. Defaults
to an object named 'SNOMED' in the global environment}

\item{...}{additional arguments to send to grepl if using
regular expression matching}
}
\value{
a vector of unique SNOMED CT concept IDs in integer64 format
}
\description{
Carries out an exact or regular expression match to
return the concept ID for a set of search terms
}
\examples{
conceptId('Heart failure', SNOMED = sampleSNOMED())
}
