% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb.R
\name{createCDB}
\alias{createCDB}
\title{Creates an environment containing CDB files}
\usage{
createCDB(
  SNOMED = getSNOMED(),
  TRANSITIVE = NULL,
  WN = NULL,
  MANUAL_SYNONYMS = NULL,
  noisy = TRUE,
  stopwords = c("the", "of", "by", "with", "to", "into", "and", "or", "at", "as",
    "and/or", "in"),
  simplify_body_part_names = TRUE,
  UMLS = NULL,
  min_length_umls_term = 5,
  umls_vocabs = c("MSH", "MEDLINEPLUS", "OMIM", "HPO", "FMA", "UWDA", "CHV", "ORPHANET",
    "NCI")
)
}
\arguments{
\item{SNOMED}{environment containing a SNOMED dictionary}

\item{TRANSITIVE}{transitive closure table, generated by
createTransitive. It is regenerated if not provided.}

\item{WN}{WordNet data.table as returned by downloadWordnet
containing WordNet data from appropriate
categories, in the format: cat (character), wordnetId (integer64),
synonyms (list), parents (list), adj (list)}

\item{MANUAL_SYNONYMS}{data.table with columns term1 and term2,
containing additional exact synonyms or abbreviations}

\item{noisy}{whether to output status messages}

\item{stopwords}{vector of stopwords}

\item{simplify_body_part_names}{Boolean, whether to add additional
body part names without preceding phrases such as 'structure of'.
Default TRUE}

\item{UMLS}{data.frame or data.table containing the MRCONSO table
with columns CUI, SAB, LAT, CODE, TTY and STR, or filepath to the
RRF format file containing this data}

\item{min_length_umls_term}{minimum length of UMLS terms to
include, as there are some short ambiguous terms that may lead
to false positives and are not useful to include. Default 5.}

\item{umls_vocabs}{which vocabularies other than SNOMED CT to draw
synonyms from. Default MSH, MEDLINEPLUS, OMIM, HPO, FMA, UWDA, CHV,
ORPHANET and NCI}
}
\value{
environment containing the following data tables: FINDINGS,
  QUAL, CAUSES, BODY, OTHERCAUSE, OTHERSEARCH, OVERLAP, TRANSITIVE,
  as well as the following source data and metadata:
  WN (wordnet table used to create the CDB), MANUAL_SYNONYMS,
  metadata (SNOMED CT version and whether UMLS_included),
  stopwords, and some SNOMEDconcept concepts for convenience:
  SCT_assoc, SCT_cause, SCT_after, SCT_dueto, SCT_findingsite
  SCT_disorder, SCT_finding, SCT_allergy
}
\description{
Extracts SNOMED CT concepts from appropriate places in the 
hierarchy to create a set of CDB files in an environment.
Uses WordNet and manual synonyms if available.
}
\examples{
# Not run
# data(MANUAL_SYNONYMS)
# WN <- downloadWordnet()
# MANUAL_SYNONYMS <- rbind(MANUAL_SYNONYMS, downloadOrphanet())
# CDBNEW <- createCDB(WN = WN, MANUAL_SYNONYMS = MANUAL_SYNONYMS)
}
\seealso{
exportMiADECDB, MANUAL_SYNONYMS, compose, decompose

Other CDB functions: 
\code{\link{addComposeLookupToCDB}()},
\code{\link{addWordnet}()},
\code{\link{batchDecompose}()},
\code{\link{downloadOrphanet}()},
\code{\link{downloadWordnet}()}
}
\concept{CDB functions}
