% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComposeLookup.R
\name{createComposeLookup}
\alias{createComposeLookup}
\title{Creates a set of lookups for SNOMED composition}
\usage{
createComposeLookup(decompositions, CDB, ...)
}
\arguments{
\item{decompositions}{filename of decompose output (read by fread) or
data.frame containing outputs of decompose function}

\item{CDB}{concept database environment, containing a table called
FINDINGS}

\item{...}{other arguments to pass to fread}
}
\value{
data.table
}
\description{
Creates composition lookup table for a set of SNOMED CT concepts
}
\examples{
# Not run

mylookup <- createComposeLookup(D)
}
\seealso{
[decompose()]
}
