% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNOMED.R
\name{createSNOMEDindices}
\alias{createSNOMEDindices}
\title{Create indices for tables in a SNOMED environment}
\usage{
createSNOMEDindices(SNOMED)
}
\arguments{
\item{SNOMED}{environment containing data.table objects: CONCEPT,
DESCRIPTION, RELATIONSHIP, STATEDRELATIONSHIP}
}
\value{
The environment with indices added to each table for
  fast searching
}
\description{
Creates relevant indices for fast searching of SNOMED CT tables
}
