% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{export}
\alias{export}
\alias{export.SNOMEDcodelist}
\title{Export a SNOMEDcodelist}
\usage{
export(x, ...)

\method{export}{SNOMEDcodelist}(x, filename = NULL, metadata = TRUE, ...)
}
\arguments{
\item{x}{SNOMEDcodelist object to export to file}

\item{...}{not used}

\item{filename}{character vector of length 1 for the file
to write to. If NULL, a filename is generated from the
codelist filename.}

\item{metadata}{whether metadata is exported (default = TRUE)}
}
\value{
invisibly returns the exported codelist
}
\description{
Writes a SNOMEDcodelist to file. If the filename is NULL,
a filename is created from the 'codelist_name' attribute.
}
\seealso{
Other SNOMEDcodelist functions: 
\code{\link{SNOMEDcodelist}()},
\code{\link{addInactiveConcepts}()},
\code{\link{expandSNOMED}()},
\code{\link{is.SNOMEDcodelist}()},
\code{\link{print.SNOMEDcodelist}()}

Other SNOMEDcodelist functions: 
\code{\link{SNOMEDcodelist}()},
\code{\link{addInactiveConcepts}()},
\code{\link{expandSNOMED}()},
\code{\link{is.SNOMEDcodelist}()},
\code{\link{print.SNOMEDcodelist}()}
}
\concept{SNOMEDcodelist functions}
