% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb.R
\name{exportMiADECDB}
\alias{exportMiADECDB}
\title{Exports CDB files for MedCAT / MiADE}
\usage{
exportMiADECDB(CDB, SNOMED, lang_refset_files, export_folderpath)
}
\arguments{
\item{CDB}{concept database environment created by creatCDB}

\item{SNOMED}{environment containing a SNOMED dictionary}

\item{lang_refset_files}{character vector of file paths to
language refset files}

\item{export_folderpath}{folder path to export to}
}
\value{
TRUE if successful
}
\description{
Uses CDB files to  format into the  SNOMED CT concepts from appropriate places in the 
hierarchy to create a set of CDB files in an environment.
Uses WordNet and manual synonyms if available.
}
\examples{
#Not run
#data(MANUAL_SYNONYMS)
#WN <- downloadWordnet()
#D <- createCDB(WN = WN, MANUAL_SYNONYMS = MANUAL_SYNONYMS)
}
