% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefset.R
\name{getRefset}
\alias{getRefset}
\title{Retrieves a Refset from the REFSET table}
\usage{
getRefset(conceptIds, SNOMED = getSNOMED())
}
\arguments{
\item{conceptIds}{character or integer64 vector of Refset SNOMED
concept IDs, or something that can be coerced to a SNOMEDconcept}

\item{SNOMED}{environment containing a SNOMED dictionary}
}
\value{
a SNOMEDconcept vector of conceptIds of members of the
  selected refset(s)
}
\description{
Retrieves a Refset from the REFSET table
}
\examples{
SNOMED <- sampleSNOMED()

getRefset(c('Renal clinical finding simple reference set',
  'Care planning activities simple reference set'))
}
