% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNOMED.R
\name{getSNOMED}
\alias{getSNOMED}
\title{Retrieves SNOMED CT dictionary from the global environment}
\usage{
getSNOMED()
}
\value{
SNOMED environment from the global environment
}
\description{
Returns an object named 'SNOMED' from the global
environment. Returns an error if no such object exists,
or if it is not an environment containing tables named
CONCEPT, RELATIONSHIP, STATEDRELATIONSHIP and DESCRIPTION.
There is no attempt to check that these tables are actually, if available. environment containing a selection of SNOMED CT
terms, their relationships and descriptions which are
provided with the package
}
\examples{
SNOMED <- sampleSNOMED()
SNOMED2 <- getSNOMED()

# To display metadata for this SNOMED CT dictionary
SNOMED2$metadata
}
