% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNOMED.R
\name{loadSNOMED}
\alias{loadSNOMED}
\title{Load SNOMED CT files from a folder(s) into R data.table objects}
\usage{
loadSNOMED(folders, active_only = TRUE)
}
\arguments{
\item{folders}{Vector of folder paths containing SNOMED CT files}

\item{active_only}{Whether to limit to current (active) SNOMED CT terms}
}
\value{
An environment containing data.table objects: CONCEPT,
  DESCRIPTION, RELATIONSHIP, STATEDRELATIONSHIP, REFSET_SIMPLE,
  REFSET_SIMPLEMAP, REFSET_EXTENDEDMAP
}
\description{
Identifies relevant SNOMED CT files from the 'Snapshot' of a 
distribution and loads them into an R environment. Files from
two folders (e.g. International and UK versions) can be loaded
together and appended.
}
\examples{
# Create a TEST environment and load the sample dictionaries
TEST <- sampleSNOMED()

# Export to temporary directory
write.table(get('CONCEPT', envir = TEST), paste0(tempdir(),
  '/_Concept_Snapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('DESCRIPTION', envir = TEST), paste0(tempdir(),
  '/_Description_Snapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('RELATIONSHIP', envir = TEST), paste0(tempdir(),
  '/_Relationship_Snapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('STATEDRELATIONSHIP', envir = TEST), paste0(tempdir(),
  '/_StatedRelationship_Snapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('REFSET', envir = TEST), paste0(tempdir(),
  '/Refset_SimpleSnapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('SIMPLEMAP', envir = TEST), paste0(tempdir(),
  '/Refset_SimpleMapSnapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)
write.table(get('EXTENDEDMAP', envir = TEST), paste0(tempdir(),
  '/Refset_ExtendedMapSnapshot.txt'), row.names = FALSE, sep = '\t',
  quote = FALSE)

# Try to import using the loadSNOMED function
TEST2 <- loadSNOMED(tempdir(), active_only = FALSE)

# Check that reimported SNOMED dictionary is the same as the original
all.equal(TEST$CONCEPT, TEST2$CONCEPT)
all.equal(TEST$DESCRIPTION, TEST2$DESCRIPTION)
all.equal(TEST$RELATIONSHIP, TEST2$RELATIONSHIP)
all.equal(TEST$STATEDRELATIONSHIP, TEST2$STATEDRELATIONSHIP)
all.equal(TEST$REFSET, TEST2$REFSET)
all.equal(TEST$SIMPLEMAP, TEST2$SIMPLEMAP)
all.equal(TEST$EXTENDEDMAP, TEST2$EXTENDEDMAP)
}
\seealso{
loadREADMAPS, CONCEPT, DESCRIPTION, RELATIONSHIP,
STATEDRELATIONSHIP, sampleSNOMED, getSNOMED
}
