% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{print.SNOMEDcodelist}
\alias{print.SNOMEDcodelist}
\title{Display a SNOMEDcodelist on screen}
\usage{
\method{print}{SNOMEDcodelist}(x, ...)
}
\arguments{
\item{x}{SNOMEDcodelist object to print to screen}

\item{...}{not used}
}
\value{
invisibly returns the codelist
}
\description{
Displays a SNOMEDcodelist on screen, including metadata.
Truncates term descriptions in order to fit within the line width.
}
\seealso{
Other SNOMEDcodelist functions: 
\code{\link{SNOMEDcodelist}()},
\code{\link{addInactiveConcepts}()},
\code{\link{expandSNOMED}()},
\code{\link{export}()},
\code{\link{is.SNOMEDcodelist}()}
}
\concept{SNOMEDcodelist functions}
