% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{print.SNOMEDconcept}
\alias{print.SNOMEDconcept}
\title{Display a SNOMEDconcept object with descriptions}
\usage{
\method{print}{SNOMEDconcept}(x, SNOMED = NULL)
}
\arguments{
\item{x}{object to check}
}
\value{
a logical vector of length one: TRUE or FALSE
}
\description{
SNOMEDconcept is an S3 class for vectors of SNOMED concept IDs
as 64-bit integers. This function checks whether the object has
the class SNOMEDconcept and is a vector of 64-bit integers.
}
\seealso{
Other SNOMEDconcept functions: 
\code{\link{SNOMEDconcept}()},
\code{\link{is.SNOMEDconcept}()}
}
\concept{SNOMEDconcept functions}
