% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{print.SNOMEDfindings}
\alias{print.SNOMEDfindings}
\title{Print method for output of 'decompose' function}
\usage{
\method{print}{SNOMEDfindings}(x, ...)
}
\description{
Print method for output of 'decompose' function
}
