% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdiagnosislist.R
\docType{package}
\name{Rdiagnosislist-package}
\alias{Rdiagnosislist}
\alias{Rdiagnosislist-package}
\title{Rdiagnosislist: A package for manipulating SNOMED CT diagnosis lists}
\description{
The Rdiagnosislist package makes it easy to load a SNOMED dictionary
into R and use the hierarchies to search for concepts and navigate
relations between concepts.
}
\author{
\strong{Maintainer}: Anoop D. Shah \email{anoop@doctors.org.uk} (\href{https://orcid.org/0000-0002-8907-5724}{ORCID})

}
\keyword{internal}
