% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdiagnosislist.R
\docType{package}
\name{Rdiagnosislist}
\alias{Rdiagnosislist}
\title{Rdiagnosislist: A package for manipulating SNOMED CT diagnosis lists}
\description{
The Rdiagnosislist provides three important categories of functionality:
Manipulating SNOMED CT code lists
Manipulating diagnosis lists
Combining existing and new diagnosis information (including matching up different records of the same diagnosis)
}
