% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{refineSNOMEDfinding}
\alias{refineSNOMEDfinding}
\title{Select more specific SNOMED CT concepts based on attributes}
\usage{
refineSNOMEDfinding(
  conceptId,
  CDB,
  composeLookup,
  attributes_conceptIds = bit64::as.integer64(0),
  due_to_conceptIds = bit64::as.integer64(0),
  without_conceptIds = bit64::as.integer64(0),
  SNOMED = getSNOMED()
)
}
\arguments{
\item{conceptId}{SNOMED CT concept to refine}

\item{CDB}{SNOMED CT concept database, as created by createCDB.
  An environment containing the following data tables: FINDINGS,
  QUAL, CAUSES, BODY, FINDINGS, OTHERSUB, OVERLAP, TRANSITIVE
composeLookup lookup table created by createComposeLookup
attributes_conceptIds SNOMED concept Ids of attributes of concept
  e.g. laterality, severity, acuteness
due_to_conceptIds SNOMED concept Ids of cause
without_conceptIds SNOMED concept Ids of conditions stated to be
  absent}

\item{SNOMED}{environment containing SNOMED CT tables}
}
\value{
a refined SNOMED concept Id
}
\description{
Finds the most specific SNOMED CT concepts that matches the
combination of a root concept and attributes. Based on a
composeLookup table
}
\examples{
# Not run
# refineSNOMEDfinding
}
