% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{SNOMEDcodelist}
\alias{SNOMEDcodelist}
\alias{as.SNOMEDcodelist}
\title{Convert a data.frame to a SNOMEDcodelist object}
\usage{
SNOMEDcodelist(x, SNOMED = get("SNOMED", envir = globalenv()))

as.SNOMEDcodelist(x, ...)
}
\arguments{
\item{x}{data.frame or data.table to convert to a SNOMEDcodelist,
with a column named 'conceptId', OR a vector of SNOMED concept IDs}

\item{SNOMED}{environment containing a SNOMED dictionary}

\item{...}{other arguments to pass to SNOMEDcodelist}
}
\value{
An object of class 'SNOMEDcodelist'
}
\description{
SNOMEDcodelist is an S3 class for lists of SNOMED codes.
It consists of conceptId and include_desc columns. The 
option to include descendants allows the creation of more succinct
SNOMED codelists.
Input is a data.frame or data.table with column names 'conceptId'
and optionally 'include_desc', which is FALSE by default, but if
TRUE then the codelist automatically includes all descendants of that
concept.
}
\details{
as.SNOMEDcodelist converts its argument into a SNOMEDcodelist but
  leaves it unchanged if it is already a SNOMEDcodelist.
}
\examples{
SNOMED <- sampleSNOMED()

my_concepts <- conceptId('Heart failure')
SNOMEDcodelist(my_concepts)
SNOMEDcodelist(data.frame(conceptId = my_concepts))
as.SNOMEDcodelist(data.frame(conceptId = my_concepts,
  include_desc = TRUE))
}
\seealso{
Other SNOMEDcodelist functions: 
\code{\link{expandSNOMED}()},
\code{\link{is.SNOMEDcodelist}()}

Other SNOMEDcodelist functions: 
\code{\link{expandSNOMED}()},
\code{\link{is.SNOMEDcodelist}()}
}
\concept{SNOMEDcodelist functions}
