% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb.R
\name{std_term}
\alias{std_term}
\title{Sample inclusion, exclusion and blacklist sets for a MiADE CDB}
\usage{
std_term(
  x,
  stopwords = c("the", "of", "by", "with", "to", "into", "and", "or", "both", "at", "as",
    "and/or", "in"),
  hyphens_to_space = FALSE,
  remove_stopwords = FALSE,
  remove_words_in_parentheses = FALSE
)
}
\arguments{
\item{x}{text to standardise}

\item{stopwords}{character vector of words to ignore or remove}

\item{hyphens_to_space}{whether to convert hyphens to spaces
(default is to remove them)}

\item{remove_stopwords}{whether to remove stopwords}

\item{remove_words_in_parentheses}{whether to remove words in
parentheses (which are usually an alternative form or explanation
for other parts of the phrase)}
}
\value{
standardised text with space at start and end, lower case
  (except for capitalised words or words with unusual
  capitalisation) and other options applied as needed
}
\description{
Returns a set of SNOMED concepts (as a SNOMEDconcept vector)
which can be used to exclude
findings in the MedCAT named entity recognition step, or blacklist
(filter out) findings from the final output.
}
\details{
\describe{
  \item{exclude_irrelevant_findings}{social history 
     (except housing problems and care needs),
     administrative statuses (except registered disabled) and
     for concept detection}
  \item{blacklist_vague_findings}{vague findings and disorders,
     intended to be used in the blacklist}
  \item{blacklist_almost_all_except_diseases}{almost all findings
     and vague disorders, intended to be used in the blacklist}
}
}
