% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComposeLookup.R
\name{createComposeLookup}
\alias{createComposeLookup}
\title{Creates a set of lookups for SNOMED composition}
\usage{
createComposeLookup(conceptIds, CDB, SNOMED = getSNOMED(), ...)
}
\arguments{
\item{conceptIds}{for which to create lookup table}

\item{CDB}{concept database environment, containing a table called
FINDINGS}

\item{SNOMED}{environment containing a SNOMED dictionary}
}
\value{
data.table
}
\description{
Creates composition lookup table for a set of SNOMED CT concepts
}
\examples{
# Not run

mylookup <- createComposeLookup(D)
}
\seealso{
[decompose()]
}
