% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{WORK IN PROGRESS}
\usage{
decompose(
  the_conceptId,
  diagnosis_text,
  CDB,
  SNOMED = getSNOMED(),
  noisy = FALSE
)
}
\arguments{
\item{CDB}{an environment containing CDB files, as created by
createCDB}

\item{SNOMED}{environment containing a SNOMED dictionary}

\item{WN}{WordNet data.table as returned by downloadWordnet
containing WordNet data from appropriate
categories, in the format: cat (character), wordnetId (integer64),
synonyms (list), parents (list), adj (list)}
}
\value{
environment containing the following data tables: FINDINGS,
  QUAL, CAUSES, BODY, FINDINGS, OTHERSUB, OVERLAP
}
\description{
Extracts SNOMED CT concepts from appropriate places in the 
hierarchy to create a set of CDB files in an environment.
Uses WordNet and manual synonyms if available.
}
\examples{
# Not run
# WORDNET <- downloadWordnet()
}
\references{
\url{https://wordnet.princeton.edu/}
}
\seealso{
[addWordnet()]
}
