% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNOMED.R
\name{exportSNOMEDenvir}
\alias{exportSNOMEDenvir}
\title{Export a SNOMED environment to a folder}
\usage{
exportSNOMEDenvir(SNOMED, folder)
}
\arguments{
\item{SNOMED}{environment containing data.table objects: CONCEPT,
DESCRIPTION, RELATIONSHIP, STATEDRELATIONSHIP, REFSET,
SIMPLEMAP, EXTENDEDMAP}

\item{folder}{path to folder where files will be written}
}
\description{
Creates tab separated files which can be reloaded with relevant indices for fast searching of SNOMED CT tables
}
\seealso{
CONCEPT, DESCRIPTION, RELATIONSHIP, STATEDRELATIONSHIP
}
