% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{htmlCodelistHierarchy}
\alias{htmlCodelistHierarchy}
\title{Export a SNOMEDcodelist hierarchy to HTML}
\usage{
htmlCodelistHierarchy(
  codelist_with_hierarchy,
  file = NULL,
  title = NULL,
  description = NULL
)
}
\arguments{
\item{codelist_with_hierarchy}{output of showCodelistHierarchy}

\item{file}{filename to export to. If NULL, no file is written}

\item{title}{title of HTML document}

\item{description}{paragraph of description text to fit within
<p></p> HTML tags}
}
\value{
a character vector containing HTML output
}
\description{
Exports a codelist with hierarchy as HTML for easy viewing.
}
\examples{
SNOMED <- sampleSNOMED()

my_concepts <- SNOMEDconcept('Heart failure')
my_codelist <- SNOMEDcodelist(data.frame(conceptId = my_concepts,
  include_desc = TRUE))
codelist_with_hierarchy <- showCodelistHierarchy(my_codelist)
htmlCodelistHierarchy(codelist_with_hierarchy)
}
\seealso{
showCodelistHierarchy
}
