% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOMEDcodelist.R
\name{is.SNOMEDcodelist}
\alias{is.SNOMEDcodelist}
\title{Check if an object is a SNOMEDcodelist}
\usage{
is.SNOMEDcodelist(x)
}
\arguments{
\item{x}{object to check}
}
\value{
a logical vector of length one: TRUE or FALSE
}
\description{
SNOMEDcodelist is an S3 class for lists of SNOMED codes.
This function checks whether the object has the class
SNOMEDcodelist. It does not check whether it contains valid data.
}
\seealso{
Other SNOMEDcodelist functions: 
\code{\link{SNOMEDcodelist}()},
\code{\link{expandSNOMED}()}
}
\concept{SNOMEDcodelist functions}
