% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNOMED.R
\name{loadSNOMED}
\alias{loadSNOMED}
\title{Load SNOMED files from a folder(s) into R data.table objects}
\usage{
loadSNOMED(folders, active_only = TRUE)
}
\arguments{
\item{active_only}{Whether to limit to current (active) SNOMED CT terms}

\item{folderpaths}{Vector of folder paths containing SNOMED CT files}
}
\value{
An environment containing data.table objects: CONCEPT,
  DESCRIPTION, RELATIONSHIP, STATEDRELATIONSHIP
}
\description{
Identifies relevant SNOMED files from a distribution and loads 
them into R. Files from two folders (e.g. International and UK
versions) can be loaded together and appended.
}
\examples{
# Load from the data folder

}
