% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MANUAL_SYNONYMS.R
\docType{data}
\name{MANUAL_SYNONYMS}
\alias{MANUAL_SYNONYMS}
\title{Sample manual synonym table to assist in creation of concept database}
\format{
An object of class \code{"data.table"}
}
\usage{
data(MANUAL_SYNONYMS)
}
\description{
Sample manual synonym table to assist in creation of concept database
}
\details{
\describe{
  \item{synonym}{ character: Synonym, in lower case if case is unimportant} 
  \item{snomed}{ character: SNOMED CT term, in lower case if case is unimportant}
  \item{bidirectional}{ boolean: whether synonym can only ever mean
    snomed. Not to be used for abbreviations or acronyms.}
}
}
\examples{
data(MANUAL_SYNONYMS)
str(MANUAL_SYNONYMS)
}
\seealso{
[addManual()]
}
\keyword{datasets}
