% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concepts.R
\name{print.SNOMEDconcept}
\alias{print.SNOMEDconcept}
\title{Display a SNOMEDconcept object with descriptions}
\usage{
\method{print}{SNOMEDconcept}(x, ...)
}
\arguments{
\item{x}{SNOMEDconcept object, or something that can be
coerced to one}

\item{...}{not required}
}
\value{
invisibly returns a character vector of the SNOMED CT
  concepts with descriptions separated by pipe (|)
}
\description{
SNOMEDconcept is an S3 class for vectors of SNOMED concept IDs
as 64-bit integers. This function checks whether the object has
the class SNOMEDconcept and is a vector of 64-bit integers.
}
\seealso{
Other SNOMEDconcept functions: 
\code{\link{as.data.frame.SNOMEDconcept}()},
\code{\link{c.SNOMEDconcept}()},
\code{\link{is.SNOMEDconcept}()},
\code{\link{union.SNOMEDconcept}()},
\code{\link{unique.SNOMEDconcept}()}
}
\concept{SNOMEDconcept functions}
