% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdiagnosislist.R
\docType{package}
\name{Rdiagnosislist}
\alias{Rdiagnosislist}
\title{Rdiagnosislist: A package for manipulating SNOMED CT diagnosis lists}
\description{
The Rdiagnosislist package makes it easy to load a SNOMED dictionary
into R and use the hierarchies to search for terms and navigate
relations between terms.
Manipulating SNOMED CT code lists
Navigating the SNOMED CT hierarchies
}
