% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{refineSNOMEDfinding}
\alias{refineSNOMEDfinding}
\title{WORK IN PROGRESS}
\usage{
refineSNOMEDfinding(
  finding,
  CDB,
  TRANSITIVE,
  SNOMED,
  composeLookup,
  single_output = TRUE
)
}
\arguments{
\item{finding}{SNOMEDfinding object}

\item{CDB}{SNOMED CT concept database, as created by createCDB.
An environment containing the following data tables: FINDINGS,
QUAL, CAUSES, BODY, FINDINGS, OTHERSUB, OVERLAP}

\item{TRANSITIVE}{transitive closure table}
}
\value{
a SNOMEDfinding object with one or more refinements of
  the SNOMED concept if possible
}
\description{
Extracts SNOMED CT concepts from appropriate places in the 
hierarchy to create a set of CDB files in an environment.
Uses WordNet and manual synonyms if available.
}
\examples{
# Not run
# refineSNOMEDfinding
}
