% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CONCEPT.R
\docType{data}
\name{SNOMED_CONCEPT}
\alias{SNOMED_CONCEPT}
\alias{CONCEPT}
\title{Sample concept table from SNOMED CT dictionary}
\format{
An object of class \code{"data.table"}
}
\usage{
data(CONCEPT)
}
\description{
A sample of the SNOMED CT concept table.
}
\details{
\itemize{
  \item{id}{ integer64: SNOMED CT conceptId (primary key)} 
  \item{moduleId}{ integer64: class of SNOMED CT concept (whether it is used for recording information or is a metadata concept)}
  \item{definitionStatusId}{ integer64: 900000000000074008 = primitive concept, 900000000000073002 = defined by conditions}
  \item{effectiveTime}{ IDate: when the concept became active}
  \item{active}{ logical: whether this concept is currently active}
}
}
\examples{
# Create a TEST environment and load the sample dictionaries
TEST <- new.env()
data(CONCEPT, envir = TEST)
data(DESCRIPTION, envir = TEST)
data(RELATIONSHIP, envir = TEST)
data(STATEDRELATIONSHIP, envir = TEST) 

# Show properties of the CONCEPT table
str(TEST$CONCEPT)
}
\keyword{datasets}
